+++ kymatio/scattering1d/utils.py
@@ -1,3 +1,7 @@
+import numpy as np
+import math
+from kymatio.scattering1d.filter_bank import scattering_filter_factory
+
 def compute_border_indices(J, i0, i1):
     """
     Computes border indices at all scales which correspond to the original
@@ -60,3 +64,67 @@ def compute_padding(J_pad, T):
         raise ValueError('Too large padding value, will lead to NaN errors')
     return pad_left, pad_right
 
+def compute_minimum_support_to_pad(T, J, Q, criterion_amplitude=1e-3,
+                                       normalize='l1', r_psi=math.sqrt(0.5),
+                                       sigma0=1e-1, alpha=5., P_max=5, eps=1e-7):
+
+
+    """
+    Computes the support to pad given the input size and the parameters of the
+    scattering transform.
+
+    Parameters
+    ----------
+    T : int
+        temporal size of the input signal
+    J : int
+        scale of the scattering
+    Q : int
+        number of wavelets per octave
+    normalize : string, optional
+        normalization type for the wavelets.
+        Only `'l2'` or `'l1'` normalizations are supported.
+        Defaults to `'l1'`
+    criterion_amplitude: float `>0` and `<1`, optional
+        Represents the numerical error which is allowed to be lost after
+        convolution and padding.
+        The larger criterion_amplitude, the smaller the padding size is.
+        Defaults to `1e-3`
+    r_psi : float, optional
+        Should be `>0` and `<1`. Controls the redundancy of the filters
+        (the larger r_psi, the larger the overlap between adjacent
+        wavelets).
+        Defaults to `sqrt(0.5)`.
+    sigma0 : float, optional
+        parameter controlling the frequential width of the
+        low-pass filter at J_scattering=0; at a an absolute J_scattering,
+        it is equal to :math:`\\frac{\\sigma_0}{2^J}`.
+        Defaults to `1e-1`.
+    alpha : float, optional
+        tolerance factor for the aliasing after subsampling.
+        The larger the alpha, the more conservative the value of maximal
+        subsampling is.
+        Defaults to `5`.
+    P_max : int, optional
+        maximal number of periods to use to make sure that the Fourier
+        transform of the filters is periodic.
+        `P_max = 5` is more than enough for double precision.
+        Defaults to `5`.
+    eps : float, optional
+        required machine precision for the periodization (single
+        floating point is enough for deep learning applications).
+        Defaults to `1e-7`.
+
+    Returns
+    -------
+    min_to_pad: int
+        minimal value to pad the signal on one size to avoid any
+        boundary error.
+    """
+    J_tentative = int(np.ceil(np.log2(T)))
+    _, _, _, t_max_phi = scattering_filter_factory(
+        J_tentative, J, Q, normalize=normalize, to_torch=False,
+        max_subsampling=0, criterion_amplitude=criterion_amplitude,
+        r_psi=r_psi, sigma0=sigma0, alpha=alpha, P_max=P_max, eps=eps)
+    min_to_pad = 3 * t_max_phi
+    return min_to_pad
